backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
backports.tarfile==1.2.0
boto3==1.35.49
django-celery-beat==2.7.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
honcho==2.0.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.collections==5.1.0
packaging==24.1
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.1
redis==5.2.0
requests==2.32.3
tomli==2.0.1
